move obsolete file to the root to be removed later
