+++ hima/experiments/temporal_pooling/new/test_stats.py
@@ -15,7 +15,7 @@ from wandb.sdk.wandb_run import Run
 from hima.common.config_utils import TConfig
 from hima.common.utils import ensure_list
 from hima.experiments.temporal_pooling.new.blocks.graph import Block, Stream
-from hima.experiments.temporal_pooling.new.blocks.stats import StatsTracker
+from hima.experiments.temporal_pooling.new.stats import StatsTracker
 from hima.experiments.temporal_pooling.new.stats.metrics import (
     multiplicative_loss, DISTR_SIM_PMF
 )
